[#26] Updated documentation.
