+++ docs/README.md
@@ -5,12 +5,12 @@ actionText: Get started
 actionLink: /guide/getting-started
 ---
 
-> The Druxt.js Blocks module adds [Drupal](https://drupal.org) Blocks, Content Blocks and Block Region Vue.js components to your [Nuxt.js](https://nuxtjs.org) frontend.
+> The DruxtJS Blocks module adds [Drupal](https://drupal.org) Blocks, Content Blocks and Block Region Vue.js components to your [NuxtJS](https://nuxtjs.org) frontend.
 
 **Example:** Header region for Umami theme.
 
 ```vue live
-<druxt-block-region
+<DruxtBlockRegion
   name="header"
   theme="umami"
 />
@@ -19,13 +19,13 @@ actionLink: /guide/getting-started
 **Example:** Block Content block rendered by UUID.
 
 ```vue live
-<druxt-block
+<DruxtBlock
   uuid="baefa4d3-9517-4413-8b9e-975c8affb8ac"
 />
 ```
 
-## Druxt.js
+## DruxtJS
 
-Druxt.js is a suite of modules to connect a [Drupal](https://drupal.org) JSON:API backend to a [Nuxt.js](https://nuxtjs.org) frontend.
+DruxtJS is a suite of modules to connect a [Drupal](https://drupal.org) JSON:API backend to a [NuxtJS](https://nuxtjs.org) frontend.
 
 Find out more at [http://druxtjs.org](http://druxtjs.org)
