+++ docs/guide/getting-started.md
@@ -1,10 +1,10 @@
 # Getting started
 
-## Druxt.js
+## DruxtJS
 
-Druxt.js is a suite of modules to connect a [Drupal](https://drupal.org) JSON:API backend to a [Nuxt.js](https://nuxtjs.org) frontend.
+DruxtJS is a suite of modules to connect a [Drupal](https://drupal.org) JSON:API backend to a [NuxtJS](https://nuxtjs.org) frontend.
 
-This module is included in the core [Druxt.js](http://druxtjs.org) project, the below instructions are only required if installing this module standalone.
+This module is included in the core [DruxtJS](http://druxtjs.org) project, the below instructions are only required if installing this module standalone.
 
 ## Quickstart
 
@@ -20,7 +20,7 @@ This module is included in the core [Druxt.js](http://druxtjs.org) project, the 
         'druxt-blocks'
       ],
       druxt: {
-        baseUrl: 'https://example.com'
+        baseUrl: 'https://demo-api.druxtjs.org'
       }
     }
     ```
@@ -29,9 +29,9 @@ This module is included in the core [Druxt.js](http://druxtjs.org) project, the 
 
 ### Backend
 - [Drupal](https://drupal.org) JSON:API backend.
-- [Druxt.js](https://www.drupal.org/project/druxt) Drupal module.
+- [DruxtJS](https://www.drupal.org/project/druxt) Drupal module.
 
 ### Frontend
-- [Nuxt.js](https://nuxtjs.org) frontend.
-- [Druxt.js Entity](https://entity.druxtjs.org/)
-- [Druxt.js Router](https://router.druxtjs.org/)
+- [NuxtJS](https://nuxtjs.org) frontend.
+- [DruxtJS Entity](https://entity.druxtjs.org/)
+- [DruxtJS Router](https://router.druxtjs.org/)
