--- src/__tests__/index.test.js
@@ -1,16 +1,0 @@
-import DruxtModule from '..'
-
-const mock = {
-  addPlugin: jest.fn(),
-  DruxtModule
-}
-
-test('Nuxt module', () => {
-  expect(() => { mock.DruxtModule() }).toThrow('Druxt settings missing.')
-
-  mock.options = {
-    druxt: {}
-  }
-  mock.DruxtModule()
-  expect(mock.addPlugin).toHaveBeenCalled()
-})
